fix: ammended user form validation and deletion conditions
