+++ packages/ui/src/index.ts
@@ -83,6 +83,8 @@ export {
 export { Label } from './components/ui/label';
 export { Checkbox } from './components/ui/checkbox';
 
+export { Tooltip, TooltipTrigger, TooltipContent, TooltipProvider } from './components/ui/tooltip';
+
 // Data Table Components
 export { DataTable, TableToolbar } from './components/data-table';
 export type { Column, DataTableProps, Pagination, ColumnCellType } from './components/data-table';
